working on data chunking and encryption

+++ main/dfs/data/include/Chunk.hpp
@@ -4,7 +4,7 @@
 
 class Chunk {
 public:
-  Chunk(int chunkNumber, int chunkSize, const std::string &chunkName)
+  Chunk(int chunkNumber, size_t chunkSize, const std::string &chunkName)
     : chunkNumber(chunkNumber), chunkSize(chunkSize), chunkName(chunkName) {}
 
   int getChunkNumber() { return this->chunkNumber; }
